Bolts Finished (for now...)
separated the spell objects folder into effects (trails, particles) and projectiles (probably will be renamed at some point to allow for stuff like Spikes and Barriers.)

Also make a little trail for putting behind the bolt projectiles.

I'm gonna call the Bolt Suffix done for now. I've done the White & Black secondary Bolt effects (add a light & pass through terrain respectively), but the other ones require that the Flammability backend is done - which is probably this weekend's job and the next major thing to implement. After that, we can get on with Strider and Transmission.